Move exodide_wheel class to exodide_wheel.py
